@@ -1,43 +1,52 @@
 module ariga.io/ogent
 
-go 1.17
+go 1.18
 
 require (
-	entgo.io/contrib v0.2.1-0.20220429104139-20a58de87e74
-	entgo.io/ent v0.10.2-0.20220407165443-5dd5acc08d0e
-	github.com/go-faster/errors v0.5.0
-	github.com/go-faster/jx v0.33.0
+	entgo.io/contrib v0.3.3
+	entgo.io/ent v0.11.4-0.20221001062602-1029a2d3ba2a
+	github.com/go-faster/errors v0.6.1
+	github.com/go-faster/jx v0.40.0
 	github.com/google/uuid v1.3.0
-	github.com/mattn/go-sqlite3 v1.14.10
-	github.com/ogen-go/ogen v0.18.1
+	github.com/mattn/go-sqlite3 v1.14.15
+	github.com/ogen-go/ogen v0.52.1
 	github.com/stoewer/go-strcase v1.2.0
-	github.com/stretchr/testify v1.7.1
-	go.opentelemetry.io/otel v1.6.3
-	go.opentelemetry.io/otel/metric v0.27.0
-	go.opentelemetry.io/otel/trace v1.6.3
+	github.com/stretchr/testify v1.8.0
+	go.opentelemetry.io/otel v1.11.0
+	go.opentelemetry.io/otel/metric v0.32.3
+	go.opentelemetry.io/otel/trace v1.11.0
+	go.uber.org/multierr v1.8.0
 )
 
 require (
-	ariga.io/atlas v0.3.8-0.20220324144249-d5d77d7dadfa // indirect
+	ariga.io/atlas v0.7.2-0.20220927111110-867ee0cca56a // indirect
 	github.com/agext/levenshtein v1.2.3 // indirect
 	github.com/apparentlymart/go-textseg/v13 v13.0.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/fatih/color v1.13.0 // indirect
 	github.com/ghodss/yaml v1.0.0 // indirect
+	github.com/go-faster/yamlx v0.3.6 // indirect
 	github.com/go-logr/logr v1.2.3 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/go-openapi/inflect v0.19.0 // indirect
-	github.com/google/go-cmp v0.5.7 // indirect
-	github.com/hashicorp/hcl/v2 v2.11.1 // indirect
+	github.com/google/go-cmp v0.5.8 // indirect
+	github.com/hashicorp/hcl/v2 v2.13.0 // indirect
+	github.com/mattn/go-colorable v0.1.12 // indirect
+	github.com/mattn/go-isatty v0.0.14 // indirect
 	github.com/mitchellh/go-wordwrap v1.0.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/segmentio/asm v1.1.3 // indirect
+	github.com/segmentio/asm v1.2.0 // indirect
 	github.com/sergi/go-diff v1.1.0 // indirect
 	github.com/zclconf/go-cty v1.10.0 // indirect
-	golang.org/x/mod v0.6.0-dev.0.20220106191415-9b9b3d81d5e3 // indirect
-	golang.org/x/sys v0.0.0-20220227234510-4e6760a101f9 // indirect
-	golang.org/x/text v0.3.7 // indirect
-	golang.org/x/tools v0.1.10 // indirect
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	go.uber.org/atomic v1.9.0 // indirect
+	go.uber.org/zap v1.23.0 // indirect
+	golang.org/x/exp v0.0.0-20221006183845-316c7553db56 // indirect
+	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
+	golang.org/x/net v0.0.0-20221004154528-8021a29435af // indirect
+	golang.org/x/sync v0.0.0-20220929204114-8fcdb60fdcc0 // indirect
+	golang.org/x/sys v0.0.0-20220919091848-fb04ddd9f9c8 // indirect
+	golang.org/x/text v0.3.8 // indirect
+	golang.org/x/tools v0.1.13-0.20220819182638-587a15310bdd // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
