@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package ent
 
@@ -58,8 +58,8 @@ type AllTypes struct {
 }
 
 // scanValues returns the types for scanning values from sql.Rows.
-func (*AllTypes) scanValues(columns []string) ([]interface{}, error) {
-	values := make([]interface{}, len(columns))
+func (*AllTypes) scanValues(columns []string) ([]any, error) {
+	values := make([]any, len(columns))
 	for i := range columns {
 		switch columns[i] {
 		case alltypes.FieldBytes:
@@ -85,7 +85,7 @@ func (*AllTypes) scanValues(columns []string) ([]interface{}, error) {
 
 // assignValues assigns the values that were returned from sql.Rows (after scanning)
 // to the AllTypes fields.
-func (at *AllTypes) assignValues(columns []string, values []interface{}) error {
+func (at *AllTypes) assignValues(columns []string, values []any) error {
 	if m, n := len(values), len(columns); m < n {
 		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
 	}
@@ -226,11 +226,11 @@ func (at *AllTypes) Update() *AllTypesUpdateOne {
 // Unwrap unwraps the AllTypes entity that was returned from a transaction after it was closed,
 // so that all future queries will be executed through the driver which created the transaction.
 func (at *AllTypes) Unwrap() *AllTypes {
-	tx, ok := at.config.driver.(*txDriver)
+	_tx, ok := at.config.driver.(*txDriver)
 	if !ok {
 		panic("ent: AllTypes is not a transactional entity")
 	}
-	at.config.driver = tx.drv
+	at.config.driver = _tx.drv
 	return at
 }
 
@@ -238,44 +238,62 @@ func (at *AllTypes) Unwrap() *AllTypes {
 func (at *AllTypes) String() string {
 	var builder strings.Builder
 	builder.WriteString("AllTypes(")
-	builder.WriteString(fmt.Sprintf("id=%v", at.ID))
-	builder.WriteString(", int=")
+	builder.WriteString(fmt.Sprintf("id=%v, ", at.ID))
+	builder.WriteString("int=")
 	builder.WriteString(fmt.Sprintf("%v", at.Int))
-	builder.WriteString(", int8=")
+	builder.WriteString(", ")
+	builder.WriteString("int8=")
 	builder.WriteString(fmt.Sprintf("%v", at.Int8))
-	builder.WriteString(", int16=")
+	builder.WriteString(", ")
+	builder.WriteString("int16=")
 	builder.WriteString(fmt.Sprintf("%v", at.Int16))
-	builder.WriteString(", int32=")
+	builder.WriteString(", ")
+	builder.WriteString("int32=")
 	builder.WriteString(fmt.Sprintf("%v", at.Int32))
-	builder.WriteString(", int64=")
+	builder.WriteString(", ")
+	builder.WriteString("int64=")
 	builder.WriteString(fmt.Sprintf("%v", at.Int64))
-	builder.WriteString(", uint=")
+	builder.WriteString(", ")
+	builder.WriteString("uint=")
 	builder.WriteString(fmt.Sprintf("%v", at.Uint))
-	builder.WriteString(", uint8=")
+	builder.WriteString(", ")
+	builder.WriteString("uint8=")
 	builder.WriteString(fmt.Sprintf("%v", at.Uint8))
-	builder.WriteString(", uint16=")
+	builder.WriteString(", ")
+	builder.WriteString("uint16=")
 	builder.WriteString(fmt.Sprintf("%v", at.Uint16))
-	builder.WriteString(", uint32=")
+	builder.WriteString(", ")
+	builder.WriteString("uint32=")
 	builder.WriteString(fmt.Sprintf("%v", at.Uint32))
-	builder.WriteString(", uint64=")
+	builder.WriteString(", ")
+	builder.WriteString("uint64=")
 	builder.WriteString(fmt.Sprintf("%v", at.Uint64))
-	builder.WriteString(", float32=")
+	builder.WriteString(", ")
+	builder.WriteString("float32=")
 	builder.WriteString(fmt.Sprintf("%v", at.Float32))
-	builder.WriteString(", float64=")
+	builder.WriteString(", ")
+	builder.WriteString("float64=")
 	builder.WriteString(fmt.Sprintf("%v", at.Float64))
-	builder.WriteString(", string_type=")
+	builder.WriteString(", ")
+	builder.WriteString("string_type=")
 	builder.WriteString(at.StringType)
-	builder.WriteString(", bool=")
+	builder.WriteString(", ")
+	builder.WriteString("bool=")
 	builder.WriteString(fmt.Sprintf("%v", at.Bool))
-	builder.WriteString(", uuid=")
+	builder.WriteString(", ")
+	builder.WriteString("uuid=")
 	builder.WriteString(fmt.Sprintf("%v", at.UUID))
-	builder.WriteString(", time=")
+	builder.WriteString(", ")
+	builder.WriteString("time=")
 	builder.WriteString(at.Time.Format(time.ANSIC))
-	builder.WriteString(", text=")
+	builder.WriteString(", ")
+	builder.WriteString("text=")
 	builder.WriteString(at.Text)
-	builder.WriteString(", state=")
+	builder.WriteString(", ")
+	builder.WriteString("state=")
 	builder.WriteString(fmt.Sprintf("%v", at.State))
-	builder.WriteString(", bytes=")
+	builder.WriteString(", ")
+	builder.WriteString("bytes=")
 	builder.WriteString(fmt.Sprintf("%v", at.Bytes))
 	builder.WriteByte(')')
 	return builder.String()
