@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package ent
 
@@ -261,7 +261,6 @@ func (atq *AllTypesQuery) Clone() *AllTypesQuery {
 //		GroupBy(alltypes.FieldInt).
 //		Aggregate(ent.Count()).
 //		Scan(ctx, &v)
-//
 func (atq *AllTypesQuery) GroupBy(field string, fields ...string) *AllTypesGroupBy {
 	grbuild := &AllTypesGroupBy{config: atq.config}
 	grbuild.fields = append([]string{field}, fields...)
@@ -288,7 +287,6 @@ func (atq *AllTypesQuery) GroupBy(field string, fields ...string) *AllTypesGroup
 //	client.AllTypes.Query().
 //		Select(alltypes.FieldInt).
 //		Scan(ctx, &v)
-//
 func (atq *AllTypesQuery) Select(fields ...string) *AllTypesSelect {
 	atq.fields = append(atq.fields, fields...)
 	selbuild := &AllTypesSelect{AllTypesQuery: atq}
@@ -318,10 +316,10 @@ func (atq *AllTypesQuery) sqlAll(ctx context.Context, hooks ...queryHook) ([]*Al
 		nodes = []*AllTypes{}
 		_spec = atq.querySpec()
 	)
-	_spec.ScanValues = func(columns []string) ([]interface{}, error) {
+	_spec.ScanValues = func(columns []string) ([]any, error) {
 		return (*AllTypes).scanValues(nil, columns)
 	}
-	_spec.Assign = func(columns []string, values []interface{}) error {
+	_spec.Assign = func(columns []string, values []any) error {
 		node := &AllTypes{config: atq.config}
 		nodes = append(nodes, node)
 		return node.assignValues(columns, values)
@@ -348,11 +346,14 @@ func (atq *AllTypesQuery) sqlCount(ctx context.Context) (int, error) {
 }
 
 func (atq *AllTypesQuery) sqlExist(ctx context.Context) (bool, error) {
-	n, err := atq.sqlCount(ctx)
-	if err != nil {
+	switch _, err := atq.FirstID(ctx); {
+	case IsNotFound(err):
+		return false, nil
+	case err != nil:
 		return false, fmt.Errorf("ent: check existence: %w", err)
+	default:
+		return true, nil
 	}
-	return n > 0, nil
 }
 
 func (atq *AllTypesQuery) querySpec() *sqlgraph.QuerySpec {
@@ -453,7 +454,7 @@ func (atgb *AllTypesGroupBy) Aggregate(fns ...AggregateFunc) *AllTypesGroupBy {
 }
 
 // Scan applies the group-by query and scans the result into the given value.
-func (atgb *AllTypesGroupBy) Scan(ctx context.Context, v interface{}) error {
+func (atgb *AllTypesGroupBy) Scan(ctx context.Context, v any) error {
 	query, err := atgb.path(ctx)
 	if err != nil {
 		return err
@@ -462,7 +463,7 @@ func (atgb *AllTypesGroupBy) Scan(ctx context.Context, v interface{}) error {
 	return atgb.sqlScan(ctx, v)
 }
 
-func (atgb *AllTypesGroupBy) sqlScan(ctx context.Context, v interface{}) error {
+func (atgb *AllTypesGroupBy) sqlScan(ctx context.Context, v any) error {
 	for _, f := range atgb.fields {
 		if !alltypes.ValidColumn(f) {
 			return &ValidationError{Name: f, err: fmt.Errorf("invalid field %q for group-by", f)}
@@ -509,7 +510,7 @@ type AllTypesSelect struct {
 }
 
 // Scan applies the selector query and scans the result into the given value.
-func (ats *AllTypesSelect) Scan(ctx context.Context, v interface{}) error {
+func (ats *AllTypesSelect) Scan(ctx context.Context, v any) error {
 	if err := ats.prepareQuery(ctx); err != nil {
 		return err
 	}
@@ -517,7 +518,7 @@ func (ats *AllTypesSelect) Scan(ctx context.Context, v interface{}) error {
 	return ats.sqlScan(ctx, v)
 }
 
-func (ats *AllTypesSelect) sqlScan(ctx context.Context, v interface{}) error {
+func (ats *AllTypesSelect) sqlScan(ctx context.Context, v any) error {
 	rows := &sql.Rows{}
 	query, args := ats.sql.Query()
 	if err := ats.driver.Query(ctx, query, args, rows); err != nil {
