@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package ent
 
@@ -43,8 +43,8 @@ func (e CategoryEdges) PetsOrErr() ([]*Pet, error) {
 }
 
 // scanValues returns the types for scanning values from sql.Rows.
-func (*Category) scanValues(columns []string) ([]interface{}, error) {
-	values := make([]interface{}, len(columns))
+func (*Category) scanValues(columns []string) ([]any, error) {
+	values := make([]any, len(columns))
 	for i := range columns {
 		switch columns[i] {
 		case category.FieldID:
@@ -60,7 +60,7 @@ func (*Category) scanValues(columns []string) ([]interface{}, error) {
 
 // assignValues assigns the values that were returned from sql.Rows (after scanning)
 // to the Category fields.
-func (c *Category) assignValues(columns []string, values []interface{}) error {
+func (c *Category) assignValues(columns []string, values []any) error {
 	if m, n := len(values), len(columns); m < n {
 		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
 	}
@@ -104,11 +104,11 @@ func (c *Category) Update() *CategoryUpdateOne {
 // Unwrap unwraps the Category entity that was returned from a transaction after it was closed,
 // so that all future queries will be executed through the driver which created the transaction.
 func (c *Category) Unwrap() *Category {
-	tx, ok := c.config.driver.(*txDriver)
+	_tx, ok := c.config.driver.(*txDriver)
 	if !ok {
 		panic("ent: Category is not a transactional entity")
 	}
-	c.config.driver = tx.drv
+	c.config.driver = _tx.drv
 	return c
 }
 
@@ -116,10 +116,11 @@ func (c *Category) Unwrap() *Category {
 func (c *Category) String() string {
 	var builder strings.Builder
 	builder.WriteString("Category(")
-	builder.WriteString(fmt.Sprintf("id=%v", c.ID))
-	builder.WriteString(", name=")
+	builder.WriteString(fmt.Sprintf("id=%v, ", c.ID))
+	builder.WriteString("name=")
 	builder.WriteString(c.Name)
-	builder.WriteString(", readonly=")
+	builder.WriteString(", ")
+	builder.WriteString("readonly=")
 	builder.WriteString(c.Readonly)
 	builder.WriteByte(')')
 	return builder.String()
