@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package ent
 
@@ -60,8 +60,7 @@ func (e PetEdges) CategoriesOrErr() ([]*Category, error) {
 func (e PetEdges) OwnerOrErr() (*User, error) {
 	if e.loadedTypes[1] {
 		if e.Owner == nil {
-			// The edge owner was loaded in eager-loading,
-			// but was not found.
+			// Edge was loaded but was not found.
 			return nil, &NotFoundError{label: user.Label}
 		}
 		return e.Owner, nil
@@ -79,8 +78,8 @@ func (e PetEdges) FriendsOrErr() ([]*Pet, error) {
 }
 
 // scanValues returns the types for scanning values from sql.Rows.
-func (*Pet) scanValues(columns []string) ([]interface{}, error) {
-	values := make([]interface{}, len(columns))
+func (*Pet) scanValues(columns []string) ([]any, error) {
+	values := make([]any, len(columns))
 	for i := range columns {
 		switch columns[i] {
 		case pet.FieldTagID:
@@ -102,7 +101,7 @@ func (*Pet) scanValues(columns []string) ([]interface{}, error) {
 
 // assignValues assigns the values that were returned from sql.Rows (after scanning)
 // to the Pet fields.
-func (pe *Pet) assignValues(columns []string, values []interface{}) error {
+func (pe *Pet) assignValues(columns []string, values []any) error {
 	if m, n := len(values), len(columns); m < n {
 		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
 	}
@@ -182,11 +181,11 @@ func (pe *Pet) Update() *PetUpdateOne {
 // Unwrap unwraps the Pet entity that was returned from a transaction after it was closed,
 // so that all future queries will be executed through the driver which created the transaction.
 func (pe *Pet) Unwrap() *Pet {
-	tx, ok := pe.config.driver.(*txDriver)
+	_tx, ok := pe.config.driver.(*txDriver)
 	if !ok {
 		panic("ent: Pet is not a transactional entity")
 	}
-	pe.config.driver = tx.drv
+	pe.config.driver = _tx.drv
 	return pe
 }
 
@@ -194,17 +193,21 @@ func (pe *Pet) Unwrap() *Pet {
 func (pe *Pet) String() string {
 	var builder strings.Builder
 	builder.WriteString("Pet(")
-	builder.WriteString(fmt.Sprintf("id=%v", pe.ID))
-	builder.WriteString(", name=")
+	builder.WriteString(fmt.Sprintf("id=%v, ", pe.ID))
+	builder.WriteString("name=")
 	builder.WriteString(pe.Name)
-	builder.WriteString(", weight=")
+	builder.WriteString(", ")
+	builder.WriteString("weight=")
 	builder.WriteString(fmt.Sprintf("%v", pe.Weight))
-	builder.WriteString(", birthday=")
+	builder.WriteString(", ")
+	builder.WriteString("birthday=")
 	builder.WriteString(pe.Birthday.Format(time.ANSIC))
-	builder.WriteString(", tag_id=")
+	builder.WriteString(", ")
+	builder.WriteString("tag_id=")
 	builder.WriteString(fmt.Sprintf("%v", pe.TagID))
+	builder.WriteString(", ")
 	if v := pe.Height; v != nil {
-		builder.WriteString(", height=")
+		builder.WriteString("height=")
 		builder.WriteString(fmt.Sprintf("%v", *v))
 	}
 	builder.WriteByte(')')
