@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package ent
 
@@ -59,8 +59,7 @@ func (e UserEdges) PetsOrErr() ([]*Pet, error) {
 func (e UserEdges) BestFriendOrErr() (*User, error) {
 	if e.loadedTypes[1] {
 		if e.BestFriend == nil {
-			// The edge best_friend was loaded in eager-loading,
-			// but was not found.
+			// Edge was loaded but was not found.
 			return nil, &NotFoundError{label: user.Label}
 		}
 		return e.BestFriend, nil
@@ -69,8 +68,8 @@ func (e UserEdges) BestFriendOrErr() (*User, error) {
 }
 
 // scanValues returns the types for scanning values from sql.Rows.
-func (*User) scanValues(columns []string) ([]interface{}, error) {
-	values := make([]interface{}, len(columns))
+func (*User) scanValues(columns []string) ([]any, error) {
+	values := make([]any, len(columns))
 	for i := range columns {
 		switch columns[i] {
 		case user.FieldID, user.FieldAge, user.FieldHeight:
@@ -88,7 +87,7 @@ func (*User) scanValues(columns []string) ([]interface{}, error) {
 
 // assignValues assigns the values that were returned from sql.Rows (after scanning)
 // to the User fields.
-func (u *User) assignValues(columns []string, values []interface{}) error {
+func (u *User) assignValues(columns []string, values []any) error {
 	if m, n := len(values), len(columns); m < n {
 		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
 	}
@@ -168,11 +167,11 @@ func (u *User) Update() *UserUpdateOne {
 // Unwrap unwraps the User entity that was returned from a transaction after it was closed,
 // so that all future queries will be executed through the driver which created the transaction.
 func (u *User) Unwrap() *User {
-	tx, ok := u.config.driver.(*txDriver)
+	_tx, ok := u.config.driver.(*txDriver)
 	if !ok {
 		panic("ent: User is not a transactional entity")
 	}
-	u.config.driver = tx.drv
+	u.config.driver = _tx.drv
 	return u
 }
 
@@ -180,18 +179,23 @@ func (u *User) Unwrap() *User {
 func (u *User) String() string {
 	var builder strings.Builder
 	builder.WriteString("User(")
-	builder.WriteString(fmt.Sprintf("id=%v", u.ID))
-	builder.WriteString(", name=")
+	builder.WriteString(fmt.Sprintf("id=%v, ", u.ID))
+	builder.WriteString("name=")
 	builder.WriteString(u.Name)
-	builder.WriteString(", age=")
+	builder.WriteString(", ")
+	builder.WriteString("age=")
 	builder.WriteString(fmt.Sprintf("%v", u.Age))
-	builder.WriteString(", height=")
+	builder.WriteString(", ")
+	builder.WriteString("height=")
 	builder.WriteString(fmt.Sprintf("%v", u.Height))
-	builder.WriteString(", favorite_cat_breed=")
+	builder.WriteString(", ")
+	builder.WriteString("favorite_cat_breed=")
 	builder.WriteString(fmt.Sprintf("%v", u.FavoriteCatBreed))
-	builder.WriteString(", favorite_dog_breed=")
+	builder.WriteString(", ")
+	builder.WriteString("favorite_dog_breed=")
 	builder.WriteString(fmt.Sprintf("%v", u.FavoriteDogBreed))
-	builder.WriteString(", favorite_fish_breed=")
+	builder.WriteString(", ")
+	builder.WriteString("favorite_fish_breed=")
 	builder.WriteString(fmt.Sprintf("%v", u.FavoriteFishBreed))
 	builder.WriteByte(')')
 	return builder.String()
