@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package ent
 
@@ -180,9 +180,15 @@ func (atc *AllTypesCreate) Save(ctx context.Context) (*AllTypes, error) {
 			}
 			mut = atc.hooks[i](mut)
 		}
-		if _, err := mut.Mutate(ctx, atc.mutation); err != nil {
+		v, err := mut.Mutate(ctx, atc.mutation)
+		if err != nil {
 			return nil, err
 		}
+		nv, ok := v.(*AllTypes)
+		if !ok {
+			return nil, fmt.Errorf("unexpected node type %T returned from AllTypesMutation", v)
+		}
+		node = nv
 	}
 	return node, err
 }
@@ -280,7 +286,7 @@ func (atc *AllTypesCreate) sqlSave(ctx context.Context) (*AllTypes, error) {
 	_node, _spec := atc.createSpec()
 	if err := sqlgraph.CreateNode(ctx, atc.driver, _spec); err != nil {
 		if sqlgraph.IsConstraintError(err) {
-			err = &ConstraintError{err.Error(), err}
+			err = &ConstraintError{msg: err.Error(), wrap: err}
 		}
 		return nil, err
 	}
@@ -493,7 +499,7 @@ func (atcb *AllTypesCreateBulk) Save(ctx context.Context) ([]*AllTypes, error) {
 					// Invoke the actual operation on the latest mutation in the chain.
 					if err = sqlgraph.BatchCreate(ctx, atcb.driver, spec); err != nil {
 						if sqlgraph.IsConstraintError(err) {
-							err = &ConstraintError{err.Error(), err}
+							err = &ConstraintError{msg: err.Error(), wrap: err}
 						}
 					}
 				}
@@ -501,11 +507,11 @@ func (atcb *AllTypesCreateBulk) Save(ctx context.Context) ([]*AllTypes, error) {
 					return nil, err
 				}
 				mutation.id = &nodes[i].ID
-				mutation.done = true
 				if specs[i].ID.Value != nil && nodes[i].ID == 0 {
 					id := specs[i].ID.Value.(int64)
 					nodes[i].ID = uint32(id)
 				}
+				mutation.done = true
 				return nodes[i], nil
 			})
 			for i := len(builder.hooks) - 1; i >= 0; i-- {
