@@ -1,4 +1,4 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package ent
 
@@ -541,7 +541,7 @@ func (atu *AllTypesUpdate) sqlSave(ctx context.Context) (n int, err error) {
 		if _, ok := err.(*sqlgraph.NotFoundError); ok {
 			err = &NotFoundError{alltypes.Label}
 		} else if sqlgraph.IsConstraintError(err) {
-			err = &ConstraintError{err.Error(), err}
+			err = &ConstraintError{msg: err.Error(), wrap: err}
 		}
 		return 0, err
 	}
@@ -797,9 +797,15 @@ func (atuo *AllTypesUpdateOne) Save(ctx context.Context) (*AllTypes, error) {
 			}
 			mut = atuo.hooks[i](mut)
 		}
-		if _, err := mut.Mutate(ctx, atuo.mutation); err != nil {
+		v, err := mut.Mutate(ctx, atuo.mutation)
+		if err != nil {
 			return nil, err
 		}
+		nv, ok := v.(*AllTypes)
+		if !ok {
+			return nil, fmt.Errorf("unexpected node type %T returned from AllTypesMutation", v)
+		}
+		node = nv
 	}
 	return node, err
 }
@@ -1095,7 +1101,7 @@ func (atuo *AllTypesUpdateOne) sqlSave(ctx context.Context) (_node *AllTypes, er
 		if _, ok := err.(*sqlgraph.NotFoundError); ok {
 			err = &NotFoundError{alltypes.Label}
 		} else if sqlgraph.IsConstraintError(err) {
-			err = &ConstraintError{err.Error(), err}
+			err = &ConstraintError{msg: err.Error(), wrap: err}
 		}
 		return nil, err
 	}
