@@ -1,10 +1,10 @@
-// Code generated by entc, DO NOT EDIT.
+// Code generated by ent, DO NOT EDIT.
 
 package runtime
 
 // The schema-stitching logic is generated in ariga.io/ogent/internal/integration/ogent/ent/runtime.go
 
 const (
-	Version = "v0.10.2-0.20220407165443-5dd5acc08d0e"           // Version of ent codegen.
-	Sum     = "h1:pyi1r2QHcaiudAGrZpo+nKM89F4r0pHhnuk3nb0ueXM=" // Sum of ent codegen.
+	Version = "v0.11.4-0.20221001062602-1029a2d3ba2a"           // Version of ent codegen.
+	Sum     = "h1:T28WZZUdeJb7DQVQNnZkr3pasIdDrDVC41eQIV0hvTU=" // Sum of ent codegen.
 )
